backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
celery
redis
reportlab
python-dotenv
